test_project/requirements.txt
Django==4.1
djangorestframework==3.14.0
channels==4.0.0
channels_redis==4.0.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
uvicorn==0.22.0
requests==2.31.0
